add Uffd support for memory snapshots
When /snapshot/load specifies memory backend type as 'UffdOverUDS',
Firecracker doesn't handle the memory file itself anymore and expects
an external process to handle its guest memory page faults.

To do this, anonymous memory is mmapped as guest memory while keeping
the original memory regions shape. Then a Uffd is created and each
guest memory range is registered with the Uffd so that any page faults
won't be handled by the kernel, but will come up as events on the Uffd.

Firecracker then sends the memory ranges descriptions/mappings along
with the Uffd over a UnixDomainSocket specified in 'mem_backend_path'
parameter on the API call.

It is expected that on the other side there is already a process
listening for incoming connections. Once Firecracker's connection is
accepted, Firecracker sends the mappings and Uffd. The receiving
process is from now responsible for handling any pagefaults on the
Uffd.

The communication medium is a UDS, the protocol is SOCK_STREAM and the
encoding is JSON.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>