pause/resume: check vcpu response
Prior to this commit, the vcpu responses were not checked
against the expected vcpu response, for Pause and Resume.

Had to remove the functions that deduplicated the checks
for pause/resume event broadcasting as response equality
cannot be easily implemented, nor is it really needed.
A match expression on the type is enough.

Signed-off-by: alindima <alindima@amazon.com>