tests/snapshots: remove test_restore_no_tsc
This test does 2 things:
  * Create a snapshot using Firecracker v0.24
  * Restore that snapshot using the current build

Firecracker v0.24 is used here because the snapshot file does not
contain the TSC frequency. We do want to allow resume on the same
processor type without any TSC adjustment but since Firecracker
v0.24 is not supported anymore we need a different type of test to
do this validation.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>